Tested map view only

+++ WebServer/routes/navigation.py
@@ -6,7 +6,9 @@ from subprocess import Popen, PIPE
 import os
 import signal
 import psutil
-from ros.node import ros_node
+from ros.topics import get_map_msg
+from mongodb.db import listMaps, listGoal
+from models.model import MapName
 
 router = APIRouter()
 process = None
@@ -14,12 +16,41 @@ process = None
 @router.get("/navigation/start")
 async def start_navigation():
     global process
-    process = Popen(['ros2', 'launch', 'autoserve_navigation', 'navigation.launch.py'], preexec_fn=os.setsid)
+    if not process:
+        process = Popen(['ros2', 'launch', 'autoserve_navigation', 'navigation.launch.py'], preexec_fn=os.setsid)
     # process = Popen(['pwd'], preexec_fn=os.setsid)
     return {'Started'}
 
 @router.get("/navigation/stop")
 async def stop_naviagtion():
-    os.killpg(os.getpgid(process.pid), signal.SIGTERM)
-    process.wait()
+    global process
+    if process:
+        os.killpg(os.getpgid(process.pid), signal.SIGTERM)
+        process.wait()
+        process = None
+    process = None
     return {'Stopped'}
+
+@router.get("/navigation/list/maps")
+async def list_maps():
+    return listMaps()
+
+@router.post("/navigation/use_map")
+async def save_map_data(name: MapName):
+    global map_name
+    print('Selected Map is', name.name)
+    map_name = name.name
+    return {"message": "Map data saved successfully"}
+
+@router.get("/navigation/list/pose/{name}")
+async def list_poses(name: str):
+    print('map name', name)
+    return listGoal(name)
+
+@router.get("/navigation/current/map")
+async def get_current_map():
+    map_msg = get_map_msg()
+    print(map_msg)
+    if map_msg:
+        return map_msg
+    return None
